+++ store/listenkv/store.go
@@ -1,8 +1,7 @@
 package listenkv
 
 import (
-	"io"
-
+	"cosmossdk.io/store/cachekv"
 	"cosmossdk.io/store/types"
 )
 
@@ -132,11 +131,5 @@ func (s *Store) GetStoreType() types.StoreType {
 // CacheWrap implements the KVStore interface. It panics as a Store
 // cannot be cache wrapped.
 func (s *Store) CacheWrap() types.CacheWrap {
-	panic("cannot CacheWrap a ListenKVStore")
-}
-
-// CacheWrapWithTrace implements the KVStore interface. It panics as a
-// Store cannot be cache wrapped.
-func (s *Store) CacheWrapWithTrace(_ io.Writer, _ types.TraceContext) types.CacheWrap {
-	panic("cannot CacheWrapWithTrace a ListenKVStore")
+	return cachekv.NewStore(s)
 }