+++ store/tracekv/store.go
@@ -6,6 +6,7 @@ import (
 	"io"
 
 	"cosmossdk.io/errors"
+	"cosmossdk.io/store/cachekv"
 	"cosmossdk.io/store/types"
 )
 
@@ -164,13 +165,7 @@ func (tkv *Store) GetStoreType() types.StoreType {
 // CacheWrap implements the KVStore interface. It panics because a Store
 // cannot be branched.
 func (tkv *Store) CacheWrap() types.CacheWrap {
-	panic("cannot CacheWrap a TraceKVStore")
-}
-
-// CacheWrapWithTrace implements the KVStore interface. It panics as a
-// Store cannot be branched.
-func (tkv *Store) CacheWrapWithTrace(_ io.Writer, _ types.TraceContext) types.CacheWrap {
-	panic("cannot CacheWrapWithTrace a TraceKVStore")
+	return cachekv.NewStore(tkv)
 }
 
 // writeOperation writes a KVStore operation to the underlying io.Writer as