+++ x/bank/module.go
@@ -16,6 +16,7 @@ import (
 	corestore "cosmossdk.io/core/store"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
+	storetypes "cosmossdk.io/store/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -43,7 +44,8 @@ var (
 	_ module.HasServices         = AppModule{}
 	_ module.HasInvariants       = AppModule{}
 
-	_ appmodule.AppModule = AppModule{}
+	_ appmodule.AppModule     = AppModule{}
+	_ appmodule.HasEndBlocker = AppModule{}
 )
 
 // AppModuleBasic defines the basic application module used by the bank module.
@@ -204,6 +206,10 @@ func (am AppModule) WeightedOperations(simState module.SimulationState) []simtyp
 	)
 }
 
+func (am AppModule) EndBlock(ctx context.Context) error {
+	return am.keeper.CreditVirtualAccounts(ctx)
+}
+
 // App Wiring Setup
 
 func init() {
@@ -222,6 +228,7 @@ type ModuleInputs struct {
 	StoreService  corestore.KVStoreService
 	TStoreService corestore.TransientStoreService
 	Logger        log.Logger
+	ObjStoreKey   *storetypes.ObjectStoreKey
 
 	AccountKeeper types.AccountKeeper
 
@@ -262,6 +269,7 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 		in.Cdc,
 		in.StoreService,
 		in.TStoreService,
+		in.ObjStoreKey,
 		in.AccountKeeper,
 		blockedAddresses,
 		authority.String(),